Add CSVImportable
This adds a concern for models which can be imported from a CSV file
which encapsulates the common logic and reduces the duplication. This
also helps to ensure that the functionality works the same across the
service.
